FIX: Fix LR bug

+++ CCGameFramework/base/clib2d/c2drevolute.cpp
@@ -58,9 +58,9 @@ namespace clib {
         auto centerB = b->world();
         auto anchorB = world_anchor_b();
 
-        auto str = min(std::log2(1 + p_acc.magnitude()), 10.0) * 0.2;
+        auto str = (float) (min(std::log2(1 + p_acc.magnitude()), 10) * 0.2);
         CComPtr<ID2D1SolidColorBrush> line;
-        rt->CreateSolidColorBrush(D2D1::ColorF(1 - str, 0.2, 0.2 + str), &line);
+        rt->CreateSolidColorBrush(D2D1::ColorF(1.0f - str, 0.2f, 0.2f + str), &line);
 
         if (!a->statics) {
             rt->DrawLine(c2d_world::transform(bounds, centerA), c2d_world::transform(bounds, anchorA), line);
